hint for helm input + arbitrary resource version
