+++ cmds/ocm/commands/ocmcmds/common/inputs/types/dockermulti/spec.go
@@ -20,6 +20,7 @@ import (
 	"github.com/open-component-model/ocm/pkg/contexts/oci/cpi"
 	"github.com/open-component-model/ocm/pkg/contexts/oci/repositories/artifactset"
 	"github.com/open-component-model/ocm/pkg/contexts/oci/repositories/docker"
+	"github.com/open-component-model/ocm/pkg/contexts/ocm/accessmethods/ociartifact"
 	"github.com/open-component-model/ocm/pkg/errors"
 	"github.com/open-component-model/ocm/pkg/runtime"
 	"github.com/open-component-model/ocm/pkg/utils"
@@ -167,5 +168,5 @@ func (s *Spec) GetBlob(ctx inputs.Context, nv common.NameVersion, inputFilePath 
 	if err != nil {
 		return nil, "", err
 	}
-	return blob, ociimage.Hint(nv, nv.GetName(), s.Repository, nv.GetVersion()), nil
+	return blob, ociartifact.Hint(nv, nv.GetName(), s.Repository, nv.GetVersion()), nil
 }
